# static/swagger-ui/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Icaro API",
    "version": "1.0",
    "description": "API para autenticação, perfis, vagas e insights no sistema Icaro"
  },
  "host": "si3de-backend.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Dados do usuário logado",
        "description": "Retorna as informações do usuário logado (incluindo perfil, se existir).",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    },
    "/user/profile": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Criar ou atualizar perfil",
        "description": "Cria ou atualiza o perfil do usuário logado (apenas alunos).",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "perfil",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string"
                },
                "curso": {
                  "type": "string"
                },
                "numeroMateriasConcluidas": {
                  "type": "integer"
                },
                "turno": {
                  "type": "string"
                },
                "modalidade": {
                  "type": "string"
                },
                "craValue": {
                  "type": "number"
                },
                "interesses": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "habilidades": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "linkedin": {
                  "type": "string"
                },
                "github": {
                  "type": "string"
                },
                "portfolio": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil atualizado com sucesso"
          },
          "422": {
            "description": "Erro de validação"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Use o formato: Bearer {seu_token_jwt}"
    }
  }
}